+++ src/tests/unit/tests/reports/components/assessment-report-instance-list.test.tsx
@@ -3,7 +3,7 @@
 import { shallow } from 'enzyme';
 import * as React from 'react';
 import { AssessmentReportInstanceList } from 'reports/components/assessment-report-instance-list';
-import { AssessmentReportBuilderTestHelper } from '../../assessment-report-builder-test-helper';
+import { AssessmentReportBuilderTestHelper } from '../../DetailsView/assessment-report-builder-test-helper';
 
 describe('AssessmentReportInstanceListTest', () => {
     describe('render', () => {
