+++ src/tests/unit/tests/reports/components/assessment-summary-details.test.tsx
@@ -2,11 +2,8 @@
 // Licensed under the MIT License.
 import { shallow } from 'enzyme';
 import * as React from 'react';
-import {
-    AssessmentSummaryDetails,
-    AssessmentSummaryDetailsProps,
-} from 'reports/components/assessment-summary-details';
-import { AssessmentReportBuilderTestHelper } from '../../assessment-report-builder-test-helper';
+import { AssessmentSummaryDetails, AssessmentSummaryDetailsProps } from 'reports/components/assessment-summary-details';
+import { AssessmentReportBuilderTestHelper } from '../../DetailsView/assessment-report-builder-test-helper';
 
 describe('AssessmentSummaryDetails', () => {
     describe('render', () => {
